requirements.txt
flask>=3.0.0
requests>=2.31.0
fal-client>=0.4.0
python-dotenv>=1.0.0
Pillow>=10.0.0
beautifulsoup4>=4.12.0
